fix ERC 721 parser for token transfer table
